Web_APP/requirements.txt
matplotlib>=3.7.4
numpy>=1.24.4
pandas>=2.0.3
plotly>=5.18.0
streamlit>=1.29.0
scikit-learn>=1.3.2
Pillow>=10.1.0
codecarbon>=2.3.2